doc search changes
