+++ Cargo.toml
@@ -42,7 +42,6 @@ compiler_test_utils = { path = "compiler/test_utils" }
 slvm = { path = "vm" }
 shell = { path = "shell" }
 regex = "1"
-lazy_static = "1"
 mdbook = "0.4.37"
 
 unicode-width = "0.2"
