+++ slosh_test/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2024"
 slosh_lib = { path = "../slosh_lib" }
 slosh_test_lib = { path = "../slosh_test_lib" }
 regex = { workspace = true }
-lazy_static = { workspace = true }
 bridge_adapters = { path = "../bridge_adapters" }
 slvm = { workspace = true }
 sl-compiler = { workspace = true }
